AS-21: Refactor Residence view and move to view folder
